Removed SQLite class from ReminderDatabase file
